added logic for parser directives component type, fixed syntax for comments to make it uniform, moved getNonce into a util folder, changed property in .eslintrc.json to remove non-dangerous warnings when running tests
