+++ src/test/runTest.ts
@@ -4,18 +4,15 @@ import { runTests } from '@vscode/test-electron';
 async function main() {
 	console.log('made it through the line before try block');
 	try {
-		console.log('inside try block');
 		// The folder containing the Extension Manifest package.json
 		// Passed to `--extensionDevelopmentPath`
 		const extensionDevelopmentPath = path.resolve(__dirname, '../../');
-
-		console.log('inside try block after first var declare');
 
 		// The path to the extension test script
 		// Passed to --extensionTestsPath
 		const extensionTestsPath = path.resolve(__dirname, './suite/index');
 
-		console.log('inside try block after second var declare');
+		console.log('inside try block after var declarations');
 
 		// Download VS Code, unzip it and run the integration test
 		await runTests({
