+++ src/test/test_cases/tc_0/component/App.jsx
@@ -1,6 +1,5 @@
-// import React from "react";
 export default function App() {
     return (
         <div>This is the App.</div>
     )
-}
+};
