+++ src/test/test_cases/tc_0/index.js
@@ -1,11 +1,4 @@
-// what we are building to test for 
-// we will point to index.js and call new instance of parser passing in file
-
-// expecting: 
-// if the value of new instance is an obj
-// tree is undefined with a proper file
-
-// test case 0 - simple react app with one app component
+// Test Case 0 - Simple react app with one app component
 
 import React from 'react';
 import { createRoot } from 'react-dom/client';
