+++ src/test/vscode-environment.js
@@ -1,6 +1,7 @@
 const { TestEnvironment } = require('jest-environment-node');
 const vscode = require('vscode');
 
+// Allows for VSCode Envionrment to be extended to Jest Environment 
 class VsCodeEnvironment extends TestEnvironment {
     async setup() {
         await super.setup();
