+++ src/test/vscode.js
@@ -1,1 +1,2 @@
+// Allows access to vscode object in testing
 module.exports = global.vscode;
