+++ src/utils/getNonce.ts
@@ -6,5 +6,5 @@ export function getNonce() {
       text += possible.charAt(Math.floor(Math.random() * possible.length));
     }
     return text;
-  }
+};
